Исправления после ревью
