+++ ImageFeed/Constants.swift
@@ -12,4 +12,5 @@ enum Constants {
     static let secretKey = "bqNTwbzNWLhw3sfdpx9L1Bt4aE7TUBS4LSvveoOOUac"
     static let redirectURI = "urn:ietf:wg:oauth:2.0:oob"
     static let accessScope = "public+read_user+write_likes"
+    static let baseURL = "https://api.unsplash.com"
 }
