+++ ImageFeed/ImagesList/ImagesListLikeResult.swift
@@ -1,0 +1,11 @@
+//
+//  Untitled 2.swift
+//  ImageFeed
+//
+//  Created by N L on 8.12.24..
+//
+import Foundation
+
+struct LikeResult: Codable {
+    let photo: PhotoResult?
+}
